requirements.txt
django
djangorestframework
markdown
django-filter
python-dateutil